fixed indexedDB.js import issue
